+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@
 - #130 Closes Find prompt when opening a file
 - #129 Fixes FileSystem Move and Copy to handle empty folders
 - #135 Fixes put merge error on empty folder when it exists in destination
+- Fixes error message when undoing copy of empty folder
 - Fixes Find input error not going away after using Go To Next Find Match
 - Fixes drive not being selected when opening parent
 - Various optimizations for file write operations